+++ src/utils/hooks.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react';
+import React, { useCallback, useState, useEffect } from 'react';
 
 // eslint-disable-next-line import/prefer-default-export
 export const useWindowSize = () => {
@@ -26,3 +26,23 @@ export const useWindowSize = () => {
 
   return windowSize;
 };
+
+export function useRenderContactHelpText(enterpriseConfig) {
+  const renderContactHelpText = useCallback(
+    () => {
+      const { contactEmail } = enterpriseConfig;
+      const message = 'reach out to your organization\'s edX administrator';
+      if (contactEmail) {
+        return (
+          <a className="text-underline" href={`mailto:${contactEmail}`}>
+            {message}
+          </a>
+        );
+      }
+      return message;
+    },
+    [enterpriseConfig],
+  );
+
+  return renderContactHelpText;
+}
